Setting stage for layer between DB and VCs
- 'Product-type' objects which combine smaller DB Structs into One
- a <Type>Store object which has a more convenient and interface-based
  API for CRUD

+++ epsilonsecurity/app/models/DatabaseId.java
@@ -3,14 +3,14 @@ package models;
 import javax.annotation.Nonnull;
 
 public abstract class DatabaseId {
-    @Nonnull protected long rawId;
+    @Nonnull protected int rawId;
 
     @Nonnull
     public long getRawId() {
         return rawId;
     }
 
-    public DatabaseId(long rawID) {
+    public DatabaseId(int rawID) {
         this.rawId = rawID;
     }
 }
